move convert fns to utils
