+++ tefla/__init__.py
@@ -17,4 +17,4 @@ from . import train_seg
 from . import train_ss
 from . import trainv2
 
-__version__ = '1.8.0'
+__version__ = '1.10.0'
